Separate rendering from main
