+++ Main.java
@@ -1,4 +1,3 @@
-import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
@@ -12,10 +11,8 @@ public class Main {
 
     static public Map map = new Map();
     static public GraphicsMap gmap = new GraphicsMap();
-    static public Scoreboard sb = new Scoreboard();
     static public GameKeyListener gameKeyListener;
     static ArrayList<Notifiable> notifiableList = new ArrayList<Notifiable>();
-    static MyFrame frame = new MyFrame();
     static HashMap<String,Passive> passiveComponents = new HashMap<String,Passive>();
     static HashMap<String,Active> activeComponents = new HashMap<String,Active>();
     static ArrayList<Mechanic> mechanics = new ArrayList<Mechanic>();
@@ -23,30 +20,20 @@ public class Main {
     static int roundNums = 30;
     static int activePlayerIndex = 0;
     static Player activePlayer;
+    static Renderer renderer;
     public static void main(String[] args) {
         gameKeyListener= new GameKeyListener(gmap);
-        frame.setSize(1280,720);
-        frame.setLayout(null);
-        frame.setVisible(true);
-        frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        frame.setResizable(false);
+        renderer.frame.setSize(1280,720);
+        renderer.frame.setLayout(null);
+        renderer.frame.setVisible(true);
+        renderer.frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        renderer.frame.setResizable(false);
         setupMainMenu();
     }
 
     private static void setupMainMenu()
     {
-        JLabel picLabel;
-        Image IMG;
-        ImageIcon icon ;
-        String filePath = "RESOURCES/IMAGES/Texture2D/";
-        icon = new ImageIcon(filePath+"BG_desert.png");
-        IMG = icon.getImage();
-        picLabel = new JLabel(new ImageIcon(IMG));
-        picLabel.setOpaque(false);
-        picLabel.setBounds(0, 0, 1280, 720);
-        JButton newGameButton=new JButton("Start");
-        newGameButton.setBounds(1280/2-100,200,200, 70);
-        newGameButton.addActionListener(new ActionListener() {
+        ActionListener newGameListener = new ActionListener() {
             public void actionPerformed(ActionEvent e) {
                 try {
                     setupGame();
@@ -55,21 +42,9 @@ public class Main {
                 }
                 gameLoop();
             }
-        } );
+        };
 
-        frame.getContentPane().add(newGameButton);
-        JButton exitGameButton=new JButton("Exit");
-        exitGameButton.setBounds(1280/2-100,400,200, 70);
-        exitGameButton.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                System.exit(0);
-            }
-        } );
-        frame.getContentPane().add(exitGameButton);
-        frame.getContentPane().add(picLabel);
-        frame.setLayout(null);
-        frame.setVisible( true );
-        frame.repaint();
+        renderer.drawMainMenu(newGameListener);
     }
 
     private static void setupGame() throws FileNotFoundException {
@@ -85,64 +60,14 @@ public class Main {
             }
         }
         fr.close();
-        frame.setPumps(passiveComponents);
+        renderer.frame.setPumps(passiveComponents);
         activePlayer= mechanics.get(0);
         gameKeyListener.setActivePlayer(activePlayer);
-        frame.addKeyListener(gameKeyListener);
-        frame.requestFocus();
-        drawScene();
+        renderer.frame.addKeyListener(gameKeyListener);
+        renderer.frame.requestFocus();
+        renderer.drawScene(mechanics, activePlayer, gmap, roundNums);
     }
 
-    private static void endScreen()
-    {
-        passiveComponents.clear();
-        frame.removeKeyListener(gameKeyListener);
-        frame.getContentPane().removeAll();
-        frame.getContentPane().setBackground(Color.BLACK);
-        JLabel winTeam;
-        if(sb.getMechPoints()>sb.getSabPoints())
-        {
-            winTeam = new JLabel("Mechanics won");
-            winTeam.setBounds(1280/2-150,100,500,200);
-        }
-        else if (sb.getMechPoints()<sb.getSabPoints())
-        {
-            winTeam = new JLabel("Saboteurs won");
-            winTeam.setBounds(1280/2-150,100,500,200);
-        }
-        else
-        {
-            winTeam = new JLabel("Draw");
-            winTeam.setBounds(1280/2-60,100,500,200);
-        }
-        winTeam.setForeground(Color.WHITE);
-        winTeam.setFont(new Font("Comic Sans MS",Font.BOLD,40));
-
-        JLabel mechScore = new JLabel("Mechanics: "+sb.getMechPoints());
-        mechScore.setForeground(Color.WHITE);
-        mechScore.setFont(new Font("Comic Sans MS",Font.BOLD,40));
-        mechScore.setBounds(300,300,500,200);
-        JLabel sabScore = new JLabel("Saboteurs: "+sb.getSabPoints());
-        sabScore.setForeground(Color.WHITE);
-        sabScore.setFont(new Font("Comic Sans MS",Font.BOLD,40));
-        sabScore.setBounds(700,300,500,200);
-        frame.getContentPane().add(winTeam);
-        frame.getContentPane().add(mechScore);
-        frame.getContentPane().add(sabScore);
-        JButton exitGameButton=new JButton("Exit");
-        exitGameButton.setBounds(1280/2-100,600,200, 70);
-        exitGameButton.addActionListener(new ActionListener() {
-            public void actionPerformed(ActionEvent e) {
-                System.exit(0);
-            }
-        } );
-        frame.add(exitGameButton);
-        frame.setLayout(null);
-        frame.setVisible( true );
-        frame.repaint();
-
-        
-    }
     public static void gameLoop()
     {
             if(roundNums>0)
@@ -155,10 +80,10 @@ public class Main {
                 }
                 if (activePlayerIndex==0||activePlayerIndex==2) {
                     activePlayer = mechanics.get(activePlayerIndex/2);
-                    MechanicKeyBinds();
+                    renderer.MechanicKeyBinds();
                 } else {
                     activePlayer = saboteurs.get((activePlayerIndex/2));
-                    SaboteurKeyBinds();
+                    renderer.SaboteurKeyBinds();
                 }
                 activePlayer.resetAP();
                 gameKeyListener.setActivePlayer(activePlayer);
@@ -167,7 +92,7 @@ public class Main {
                     c.MoveWater(null);	
                 }
                 map.resetMoveWater();
-                sb.TallyPoints(map);
+                renderer.sb.TallyPoints(map);
                 int mapSize=map.getComponents().size();
                 for (Notifiable n:notifiableList) 
                 {
@@ -181,13 +106,13 @@ public class Main {
                         Main.passiveComponents.put(pii.getId(),pii);
                     }
                 }
-                drawScene();
+                renderer.drawScene(mechanics, activePlayer, gmap, roundNums);
                 new JOptionPane().showMessageDialog(null, "Uj aktiv jatekos: " + activePlayer.getId(), "JATEKOS VALTAS!", JOptionPane.WARNING_MESSAGE);
                 }
-            drawScene();
+            renderer.drawScene(mechanics, activePlayer, gmap, roundNums);
             if(roundNums <= 0)
             {
-                endScreen();
+                renderer.endScreen(passiveComponents, gameKeyListener);
             }
         }
     }
@@ -319,81 +244,5 @@ public class Main {
         int newIn = Integer.parseInt(args[1]);
         int newOut = Integer.parseInt(args[2]);
         c.SetPump(newIn-1, newOut-1);
-    }
-
-    private static void drawScene()
-    {
-        frame.getContentPane().removeAll();
-        if (mechanics.contains(activePlayer))
-            MechanicKeyBinds();
-        else
-            SaboteurKeyBinds();
-            for (int i =gmap.getComponents().size()-1; i>-1;i--){
-                frame.add(gmap.getComponents().get(i).getLabel());
-                frame.add(gmap.getComponents().get(i).getIdLabel());
-            }
-
-
-        JLabel scoreLabel = new JLabel("<html><font color='white'>SAB POINTS: "+sb.getSabPoints()+"<br>MEC POINTS: "+sb.getMechPoints()+"<br><br>ROUND: "+(30-roundNums)+"/30<br><br>AP: "+activePlayer.getAP()+"<br>"+activePlayer.getId()+"</font></html>");
-        scoreLabel.setFont(new Font("Arial",1,20));
-        scoreLabel.setBounds(frame.getWidth()-215,0,198, 220);
-        scoreLabel.setOpaque(true);
-        scoreLabel.setBackground(Color.GRAY);
-        frame.add(scoreLabel);
-        SwingUtilities.updateComponentTreeUI(frame);
-
-        JLabel picLabel;
-        Image IMG;
-        ImageIcon icon ;
-        String filePath = "RESOURCES/IMAGES/Texture2D/";
-        icon = new ImageIcon(filePath+"BG_desert.png");
-        IMG = icon.getImage();
-        picLabel = new JLabel(new ImageIcon(IMG));
-        picLabel.setOpaque(false);
-        picLabel.setBounds(0, 0, 1280, 720);
-        frame.getContentPane().add(picLabel);
-        frame.setLayout(null);
-        frame.setVisible( true );
-        frame.repaint();
-    }
-
-    private static void MechanicKeyBinds(){
-        JLabel label2 = new JLabel("<html><font color='white'>UP:..........................w<br>" +
-                "RIGHT:....................d<br>" +
-                "DOWN:....................s<br>" +
-                "LEFT:......................a<br>" +
-                "ESCAPE:................e<br>" +
-                "PASS:.....................q<br>" +
-                "SET PUMP:.............v<br>"+
-                "PUNCTURE:............p<br>"+
-                "FIX:..........................f<br>"+
-                "STICKY:..................t<br>"+
-                "COLLECT PUMP:...c<br>"+
-                "PLACE PUMP:........g<br>"+
-                "RELOCATE PIPE:...r<br>"+
-                "DETACH PIPE:.......h</font></html>");
-        label2.setFont(new Font("Arial",1,18));
-        label2.setBounds(frame.getWidth()-215,210,198, frame.getHeight()-210);
-        label2.setOpaque(true);
-        label2.setBackground(Color.BLACK);
-        frame.add(label2);
-    }
-
-    private static void SaboteurKeyBinds(){
-        JLabel label = new JLabel("<html><font color='white'>UP:..........................w<br>" +
-                "RIGHT:....................d<br>" +
-                "DOWN:....................s<br>" +
-                "LEFT:.......................a<br>" +
-                "ESCAPE:.................e<br>" +
-                "PASS:......................q<br>" +
-                "SET PUMP:..............v<br>"+
-                "PUNCTURE:............p<br>"+
-                "SLIPPERY:...............l<br>"+
-                "STICKY:...................t</font></html>");
-        label.setFont(new Font("Arial",1,18));
-        label.setBounds(frame.getWidth()-215,210,198, frame.getHeight()-210);
-        label.setOpaque(true);
-        label.setBackground(Color.BLACK);
-        frame.add(label);
     }
 }
